SR_prototyping/pyproject.toml
[project]
name = "sr-prototyping"
version = "0.1.0"
description = "Super resolution algorithm prototyping and experimentation"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "opencv-python>=4.9",
    "matplotlib>=3.8",
    "pillow>=10.0",
    "scipy>=1.12",
    "scikit-image>=0.22",
    "cairosvg>=2.7",
    "prysm>=0.21",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "forward_model"]
